Handle Bucket APIs in BucketController Part 1
